requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==24.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
docker-pycreds==0.4.0
eval_type_backport==0.2.0
filelock==3.15.1
fiftyone==0.25.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.3
idna==3.7
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
mteb==1.13.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
polars==1.4.1
portalocker==2.10.1
protobuf==5.27.3
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycocotools==2.0.8
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytrec-eval-terrier==0.5.6
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sentry-sdk==2.12.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
spacy==4.0.0.dev3
sympy==1.12.1
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.6
wget==3.2
xxhash==3.4.1
yarl==1.9.4